:recycle: remove template in GreensFunctions
Some of the greens functions heavily depends on the GSL features, such
as numerical integration or extrapolation. GSL is a C library, so it
does not have any overload nor template. It means that Real type cannot
be generalized (in an easy way). So here, templates in Green's functions
only makes the code complicated and gain almost nothing. It is better to
fix Real type as double.